Implement ItemCgpRecord
