+++ crates/macros/cgp-macro-core/src/types/cgp_data/mod.rs
@@ -1,0 +1,3 @@
+mod record;
+
+pub use record::*;
