+++ crates/macros/cgp-macro-core/src/types/mod.rs
@@ -2,6 +2,7 @@ pub mod attributes;
 pub mod blanket_trait;
 pub mod cgp_auto_getter;
 pub mod cgp_component;
+pub mod cgp_data;
 pub mod cgp_fn;
 pub mod cgp_getter;
 pub mod cgp_impl;
