+++ crates/macros/cgp-macro-lib/src/entrypoints/mod.rs
@@ -17,6 +17,7 @@ mod delegate_components;
 mod derive_build_field;
 mod derive_extract_field;
 mod derive_from_variant;
+mod derive_has_field;
 mod derive_has_fields;
 mod path;
 mod product;
@@ -42,6 +43,7 @@ pub use delegate_components::*;
 pub use derive_build_field::*;
 pub use derive_extract_field::*;
 pub use derive_from_variant::*;
+pub use derive_has_field::*;
 pub use derive_has_fields::*;
 pub use path::*;
 pub use product::*;
