+++ crates/macros/cgp-macro-lib/src/lib.rs
@@ -10,10 +10,7 @@ extern crate alloc;
 pub(crate) mod derive_builder;
 pub(crate) mod derive_extractor;
 pub(crate) mod derive_has_fields;
-pub(crate) mod field;
 
 mod entrypoints;
-
-pub use field::derive_has_field;
 
 pub use crate::entrypoints::*;
